+++ crates/cli/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- Add `parquet-max-row-group-size` parameter to commands that write parquet files with default value of `150_000` ([#846](https://github.com/stac-utils/rustac/pull/846))
+
 ## [0.1.2] - 2025-11-14
 
 Update **stac** dependency.